requirements.txt
djangorestframework
pytest-django
pytest-factoryboy